refactor(cmd-shim): move long unit-test blocks into external tests.rs
Per CODE_STYLE_GUIDE.md "Unit test file layout": once a `mod tests`
block is long enough to obscure the surrounding module, it moves to
`<parent>/tests.rs`. The four test blocks introduced for bin linking
(78-130 lines each) were past that threshold.

Moved:
- crates/cmd-shim/src/bin_resolver.rs       -> .../bin_resolver/tests.rs
- crates/cmd-shim/src/shim.rs               -> .../shim/tests.rs
- crates/cmd-shim/src/link_bins.rs          -> .../link_bins/tests.rs
- crates/package-manager/src/link_bins.rs   -> .../link_bins/tests.rs

No test logic changed.
